closes #13
Made our observed position a little bit sticky so that if we
overshoot by a small amount we just land there. Should work fine
in practice, but it's a little ugly and it MAY open up the
possibility of losing steps. I tried to make my local box skip
steps and didn't see any issues, though.